忽略因网络波动导致获取ip错误。自定义的类合并为helper.py。后续核心功能没问题将不再更新
