+++ package.json
@@ -861,20 +861,18 @@
         "name": "动态企微可信IP",
         "description": "修改企微应用可信IP，支持Srever酱等第三方通知。验证码以？结尾发送到企业微信应用",
         "labels": "消息通知",
-        "version": "1.5.2",
+        "version": "1.6.0",
         "icon": "Wecom_A.png",
         "author": "RamenRa",
         "level": 2,
         "v2": true,
         "history": {
+            "v1.6.0": "忽略因网络波动导致获取ip错误。自定义的类合并为helper.py。后续核心功能没问题将不再更新",
             "v1.5.2": "可以从指定url获取ip,修复不使用cc时cookie失效过快，v1可配置第三方为备用通知，server酱可以将文本发送到server3,二维码给服务号",
             "v1.5.1": "修复v2微信通知，可以指定微信通知ID",
             "v1.5.0": "支持企微应用通知和第Serve酱等第三方推送。按要求修改插件名称",
             "v1.4.1": "完善面板说明",
-            "v1.4.0": "修复强制更改IP时配置面板延时过长的问题。庆祝v2进入正式版，显示了一个没用的参数",
-            "v1.3.1": "修正一些逻辑判断，修改ip成功会通知一次",
-            "v1.3.0": "兼容v2，操作cookie前检查一次CookieCloud",
-            "v1.2.0": "远程命令/push_qr，立即推送一次二维码到pushplus。添加<本地扫码刷新cookie>"
+            "v1.4.0": "修复强制更改IP时配置面板延时过长的问题。庆祝v2进入正式版，显示了一个没用的参数"
         }
     },
     "SyncCookieCloud": {
